Fix: Circular dependency problem
In vk_Vertex.hpp, we include vk_Buffer as a member in order to
use copy_buffer or createBuffer.

But in vk_Buffer.hpp, we include vk_CommandBuffer in order to use
recordSingleCommand, and in vk_Buffer we include vk_Vertex to
create vertex buffer, in this case circular dependency occurs.

So we need to forward the `vk_Buffer` object in header files, and
actually include the `vk_Buffer.hpp` in .cpp file, which solves
this situation.